fix: 支持长整型企业微信应用ID

+++ weixin-java-cp/src/test/java/me/chanjar/weixin/cp/corpgroup/service/impl/WxCpCgServiceApacheHttpClientImplTest.java
@@ -57,7 +57,7 @@ public class WxCpCgServiceApacheHttpClientImplTest {
   public WxCpCorpGroupConfigStorage wxCpCorpGroupConfigStorage() {
     wxCpCorpGroupConfigStorage = new WxCpCorpGroupDefaultConfigImpl();
     wxCpCorpGroupConfigStorage.setCorpId(wxCpService.getWxCpConfigStorage().getCorpId());
-    wxCpCorpGroupConfigStorage.setAgentId(wxCpService.getWxCpConfigStorage().getAgentId());
+    wxCpCorpGroupConfigStorage.setAgentId(wxCpService.getWxCpConfigStorage().getAgentId().intValue());
     return wxCpCorpGroupConfigStorage;
   }
